remove interface in service layer
